functional code - before hacking the polyfit
some modularisatoin.. this is not using the polyfit to interpolate as yet
